Add basic support for Route53 as a backend
Table description is now stored as a value on a Route53 DNS TXT record.
Data is serialized as a "modified" JSON with single instead of double
quotes to avoid uncomfortable escaping.

Created basic infrastructure and tested with GraalVM's native image.

Slightly improved the documentation with the limitations that this
project will have due to Route53 limitations compared to Dynamo.