Add DescribeTable operation
The table definition is stored as a TXT record on a record with name the
table name. The value is a JSON-like document which describe the keys
(names, datatype and keytype).

The program now supports to convert from the request into the
Route53-stored format and viceversa. Both the CreateTable and the
DescribeTable operations now return a properly formatted (albeit just a
partial subset) of the information that the operation on real Dynamo
returns.

TODO: refactor, clean, improve the code.