Fix a bug findind RRs in Route53
Method Route53Manager.getSingleValuedResource() had a bug in the way RRs
are searched in Route53. Method .startRecordName(String) in the builder
that generates the request may return a record lexicographically higher
than the parameter passed if it is not found. Therefore, it may return
results even for records that doesn't exist, so the name of the returned
record needs to be checked.