Switch item Route53 label to a SHA-224 hash
Given the restrictions on the possible characters to use as part of a
Route53 label, we should not use directly the hash key as the value for
that label. Instead, we use the hexadecimal representation of the
SHA-224 hash of the hash key. We avoid a more compact representation to
also allow it to be generated easily from other external systems (like a
shell).