Make data serialization string-escape correct
Before this commit, dyna53 stores item data in a JSON-like serialization
format, which is (was) essentially JSON with all double quotes replaced
by single quotes. This works for simple cases but may fail when string
data contains quotes inside.

Unfortunately, the JSON provider used (Jsonb) does not support changing
the quote character. Instead, a switch to Jackson is performed, which
can be customized in this way.

Both actually co-exist, Jsonb is used for all HTTP-related (parsing
requests, generating responses) for Dynamo protocol; and Jackson is used
for reading and writing to Route53, where we use this serialization
format.