Improve create-table and put-item processes
create-table is not an idempotent operation, and should fail if the
table already exists. put-item is idempotent by default and replaces
existing values if that's the case. Code has been updated to match these
behaviors, and report errors like DynamoDB.

Significant refactor of Route53Manager code, and externalizing code that
did not really belong to this abstraction layer. Adding better
management of the async calls that return CompletableFuture.

Added more comformance to the DynamoDB error reporting on some cases.

Oh, and added a very easy to find Easter Egg. Let's see how's first ;)