Add (proper) support for values with composite PKs
This commit allows to insert values with the same hash key and different
range key, as it is expected in tables with a composite primary key.
Before, the last value was overritten (given that the default put-item
behavior is to upsert), and records are only differentiated by the name,
which is solely generated from the hash key.

Now records with the same hash key are "co-located" (as they are in
proper Dynamo, in the same partition) within the same Resource Record.
This is possible as RRs allow for multiple values (up to 400) within the
record itself.

Fixed put-item, get-item and scan operations to work properly on this
context (e.g. scan generates one result per value of every record, as if
all were different records).