Fix a bug on scan operation due to Route53 results limit
Set explicitly a limit of results from a Route53 query to the maximum
number of items there could be on a table, to avoid pagination.
Otherwise, there's a default of 100 which might produce invalid results
(e.g. scanning resources of another table and reaching the limit,
therefore not returning all results of the scanned table).